+++ src/components/styles/ScenariosList.css
@@ -1,0 +1,39 @@
+.scenario {
+  margin-top: 1.5rem;
+  background-color: grey;
+  font-size: 1.2em;
+  color: rgba(0, 0, 0, 0.6);
+  padding: 1rem;
+  border-radius: 1.5rem;
+  border: 1px solid #dddddd;
+  box-shadow: 1px 1px 4px 0px rgba(35, 41, 51, 0.2);
+}
+
+.scenario:hover {
+  border: 1px solid rgba(0, 0, 0, 0.3);
+  color: rgba(0, 0, 0, 0.8);
+  cursor: pointer;
+}
+
+.scenario > .collapse {
+  font-size: 1.0em;
+}
+
+.passedScenario {
+  background-color: rgba(45, 243, 129, 0.1);
+}
+
+.failedScenario {
+  background-color: rgba(243, 129, 45, 0.1);
+}
+
+.pendingScenario {
+  background-color: rgba(227, 239, 59, 0.1);
+}
+
+@media only screen and (max-width: 767px) {
+  .scenario {
+    word-break: break-all;
+    hyphens: auto;
+  }
+}
