+++ src/containers/scenarios.jsx
@@ -1,10 +1,10 @@
 import React, { Component } from 'react';
 import { connect } from 'react-redux';
 import _ from 'lodash';
-import { CollapsibleItem, Details, PerPageFilter, Pagination,
+import ScenariosList from '../components/scenarios_list';
+import { Details, PerPageFilter, Pagination,
   ScenarioSuccessChart, SearchScenarios } from '../components';
-import { filterScenarios, firstScenarioColumn, secondScenarioColumn,
-  slicePageScenarios, statusName } from '../helpers/scenarios_helpers';
+import { firstScenarioColumn, secondScenarioColumn } from '../helpers/scenarios_helpers';
 import { getAllScenarios } from '../actions/project_scenarios_actions';
 import styles from './styles/Scenarios.css';
 
@@ -26,83 +26,18 @@ class Scenarios extends Component {
     );
   }
 
-  constructor(state) {
-    super(state);
-    this.state = { scenarios: [], page: 1, perPage: 10, total: 1, search: [] };
-    this.setPage = this.setPage.bind(this);
-    this.setPerPage = this.setPerPage.bind(this);
-    this.setSearch = this.setSearch.bind(this);
-  }
-
-  shouldComponentUpdate(nextProps, nextState) {
-    return (
-      (!_.isEqual(nextProps, this.props)) ||
-      (!_.isEqual(nextState, this.state))
-    );
-  }
-
   componentDidMount() {
     const { xApiKey, scenarios } = this.props;
     this.props.getAllScenarios(xApiKey);
-    this.setScenarios();
-  }
-
-  componentDidUpdate() {
-    this.setScenarios();
-  }
-
-  setScenarios() {
-    const examples = _.get(this.props.scenarios, 'examples');
-    const filteredScenarios = filterScenarios(examples, this.state.search);
-    const totalPages = _.ceil(this.state.total / this.state.perPage);
-    const newState = {
-      scenarios: filteredScenarios,
-      total: filteredScenarios.length,
-    }
-    if (totalPages > 1 && totalPages < this.state.page) {
-      return this.setState(_.merge(newState, { page: totalPages }));
-    }
-    this.setState(newState);
-  }
-
-  setPage({ page }) {
-    this.setState({ page });
-  }
-
-  setPerPage({ perPage }) {
-    this.setState({ perPage });
-  }
-
-  setSearch({ search }) {
-    this.setState({ search });
-  }
-
-  renderScenarios() {
-    const { scenarios, page, perPage } = this.state;
-    if (_.isEmpty(scenarios)) {
-      return (<div className="loading">No scenarios found</div>);
-    }
-    let childKey = 0;
-    return _.map(slicePageScenarios(scenarios, page, perPage), (scenario) => {
-      childKey += 1;
-      const status = statusName(scenario.last_status);
-      return (
-        <CollapsibleItem
-          className={`${styles.scenario} ${styles[status]}`}
-          title={scenario.name}
-          details={this.constructor.renderScenarioDetails(scenario)}
-          key={childKey}
-        />
-      );
-    });
+    this.props.setScenarios();
   }
 
   render() {
-    if (_.isEmpty((_.get(this.props.scenarios, 'examples')))) {
+    if (_.isEmpty((_.get(this.props.scenariosList, 'examples')))) {
       return (<div className="loading">Loading...</div>);
     }
 
-    if (_.isEmpty(this.props.scenarios)) {
+    if (_.isEmpty(this.props.scenariosList)) {
       return (<div className="loading">Have not submitted any scenarios yet.</div>);
     }
 
@@ -115,29 +50,29 @@ class Scenarios extends Component {
           </div>
           <div className={styles.allScenariosSearch}>
             <SearchScenarios
-              search={this.state.search}
-              action={this.setSearch}
+              search={this.props.search}
+              action={this.props.setSearch}
             />
           </div>
           <div className={styles.scenariosTotal}>
-            Total Scenarios: {this.state.total}
+            Total Scenarios: {this.props.total}
           </div>
           <div className={styles.allScenarios}>
-            {this.renderScenarios()}
+            {this.props.renderScenarios(this.constructor.renderScenarioDetails)}
           </div>
           <div className={styles.scenarioListButtons}>
             <Pagination
               className={styles.scenarioPagination}
-              page={this.state.page}
-              perPage={this.state.perPage}
-              total={this.state.total}
-              action={this.setPage}
+              page={this.props.page}
+              perPage={this.props.perPage}
+              total={this.props.total}
+              action={this.props.setPage}
             />
             <PerPageFilter
-              totalCount={this.state.total}
+              totalCount={this.props.total}
               buttons={[30,10]}
-              perPage={this.state.perPage}
-              action={this.setPerPage}
+              perPage={this.props.perPage}
+              action={this.props.setPerPage}
             />
           </div>
         </div>
@@ -156,8 +91,9 @@ const mergeScenarios = (projectScenarios) => {
 
 const mapStateToProps = (state) => ({
   projects: _.keys(state.projectScenarios.scenariosList.data),
-  scenarios: mergeScenarios(state.projectScenarios.scenariosList.data),
+  scenariosList: mergeScenarios(state.projectScenarios.scenariosList.data),
   xApiKey: state.users.currentUser.xApiKey,
 });
 
-export default connect(mapStateToProps, { getAllScenarios })(Scenarios);
+const composedComponent = ScenariosList(Scenarios);
+export default connect(mapStateToProps, { getAllScenarios })(composedComponent);
