+++ src/containers/styles/Scenarios.css
@@ -38,27 +38,6 @@
   grid-column: 1 / -1;
 }
 
-.scenario {
-  margin-top: 1.5rem;
-  background-color: grey;
-  font-size: 1.2em;
-  color: rgba(0, 0, 0, 0.6);
-  padding: 1rem;
-  border-radius: 1.5rem;
-  border: 1px solid #dddddd;
-  box-shadow: 1px 1px 4px 0px rgba(35, 41, 51, 0.2);
-}
-
-.scenario:hover {
-  border: 1px solid rgba(0, 0, 0, 0.3);
-  color: rgba(0, 0, 0, 0.8);
-  cursor: pointer;
-}
-
-.scenario > .collapse {
-  font-size: 1.0em;
-}
-
 .scenarioExtendedDetails {
   display: grid;
   grid-gap: 0.5em;
@@ -95,18 +74,6 @@
   display: inline-block;
 }
 
-.passedScenario {
-  background-color: rgba(45, 243, 129, 0.1);
-}
-
-.failedScenario {
-  background-color: rgba(243, 129, 45, 0.1);
-}
-
-.pendingScenario {
-  background-color: rgba(227, 239, 59, 0.1);
-}
-
 @media only screen and (max-width: 767px) {
   .scenarios {
     grid-template-rows: 70px 20px 50px auto auto;
@@ -140,10 +107,5 @@
 
   .scenarioSecondaryDetails>div:not(:nth-child(1)) {
     display: none;
-  }
-
-  .scenario {
-    word-break: break-all;
-    hyphens: auto;
   }
 }
